Windows: refactor stats
